It's a simple Guess a Number game Coded in C. Just for practice purpose. But the game is really fun and interesting give it a try.
This is a very basic game i coded in C language Just for practice but the Game turns out to be very fun. I really enjoyed while doing work on this code .
I hope you'll enjoy it.

Working----->
This code is made with a do-while loop to make sure at least it execute once. Here we use two modules to generate a random number Player will guess  a number which will be stored in a variable name Guess. We will make sure the count of guesses in variable (no_of_guesses) . then the code will check the number guessed is same as a random number it generate, If it is not same then code will give hint to think of a larger or smaller  Number. When player Guess the correct number , they will be congratulated and One with less number of guesses will WIN. It is a simple game with a very basic code. Give it a try !!